+++ frontend/pages/index.vue
@@ -54,12 +54,15 @@
           />
       </div>
       <div class="w-full mx-auto mt-0 space-x-3 space-y-3" v-if="selectedDataSources">
-        <DataSourceQuestionsHome 
-            :data_sources="selectedDataSources" 
-            @update-content="updateTextarea" 
+        <DataSourceQuestionsHome
+            :data_sources="selectedDataSources"
+            @update-content="updateTextarea"
         />
       </div>
 
+      <div class="w-full mx-auto mt-4">
+        <RecentReports />
+      </div>
 
     </div>
 
@@ -127,6 +130,7 @@ import { useExcel } from '~/composables/useExcel';
 import { onMounted, nextTick } from 'vue';
 import Spinner from '@/components/Spinner.vue'
 import PromptBoxV2 from '~/components/prompt/PromptBoxV2.vue';
+import RecentReports from '~/components/home/RecentReports.vue';
 
 import { useCan } from '~/composables/usePermissions'
 import { KeyCode } from 'monaco-editor';